requirements.txt
anyio==3.7.1
apischema==0.16.6
attrs==23.1.0
base58==2.1.1
cachetools==4.2.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
Django==4.2.5
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.4
jsonrpcclient==4.0.3
jsonrpcserver==5.0.9
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
OSlash==0.6.3
Pillow==10.0.1
# psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyNaCl==1.5.0
pypng==0.20220715.0
qrcode==7.4.2
referencing==0.30.2
requests==2.31.0
rfc3986==1.5.0
# rpds-py==0.10.3
sniffio==1.3.0
solana==0.21.0
solana-pay-py==0.1.0
sqlparse==0.4.4
types-cachetools==4.2.10
typing-extensions==3.10.0.2
tzdata==2023.3
urllib3==1.26.7
websockets==10.4
python-dotenv==1.0.1